#3 browse() :: tested, added strict parameter
Strict (true) : The method triggers a warning and returns false on first
non-existing key, even if the wrong key is at the end of the path.

Strict (false) : The method stops on first error, but return the last
correct node found with the previous key of the path.
If the first key of the path is wrong, it returns the node itself.

i.e : $node->browse("ok/correct/wrong")
return exactly the same that $node->browse("ok/correct") (wrong is
ignored)